+++ src/types/datetime/index.d.ts
@@ -1,2 +1,3 @@
 export {DateTime} from './datetime/datetime';
 export {TimeSpan} from './timespan/timespan';
+export {DateRange} from './daterange/daterange';