+++ src/utils/validation/class.validator/class.validator.js
@@ -1053,6 +1053,7 @@ function ValidateClass(method, validationMessage) {
 }
 
 module.exports = {
+    VALIDATIONS,
     ClassValidator,
     IsDefined, IsEmail, Min, Max, Blacklist, IsEmpty, IsNotEmpty, MaxLength, MinLength, ValidateIf, Whitelist,
     Equals, NotEquals,